modularize desktop shell page
